Download species' habitat information from the IUCN RedList online database (No longer supported)

Get species' habitat information from the IUCN RedList website(https://www.iucnredlist.org/) for one or more species.

Usage

lets.iucn.ha(input, count = FALSE)

Arguments

input

Character vector with one or more species names, or an object of the PresenceAbsence class.

count

Logical, if TRUE a counting window will open.

Value

A data frame with species names in the first column and the habitats where it occurs in the remaining columns, '1' if species is present in that habitat and '0' otherwise.

Details

Note that you must be connected to the internet to use this function.

See also

Author

Bruno Vilela

Examples

if (FALSE) {
# Single species
lets.iucn.ha("Pongo pygmaeus")

# Multiple species
sp <- c("Musonycteris harrisoni", "Ailuropoda melanoleuca",
        "Cebus flavius")
lets.iucn.ha(sp)
}