Subset a PresenceAbsence object based on species names

Subset a PresenceAbsence object based on species character vector provided by the user.

Usage

lets.subsetPAM(x, names, remove.cells = TRUE)

Arguments

x

A PresenceAbsence object.

names

Character vector with species names to subset the PresenceAbsence object.

remove.cells

Logical, if TRUE the final matrix will not contain cells in the grid with a value of zero (i.e. sites with no species present).

Value

The result is an object of class PresenceAbsence subseted.

See also

Author

Bruno Vilela

Examples

if (FALSE) {
data(PAM)
# PAM before subset
plot(PAM, xlab = "Longitude", ylab = "Latitude",
     main = "Phyllomedusa species richness")

# Subset PAM to the first 20 species
PAMsub <- lets.subsetPAM(PAM, PAM[[3]][1:20])
plot(PAMsub, xlab = "Longitude", ylab = "Latitude",
     main = "Phyllomedusa species richness")
}